+++ src/components/trend-icon.tsx
@@ -1,16 +1,19 @@
 import { ComponentProps } from "react";
 
+import { Trend } from "src/graphql/resolver-types";
 import { Icon } from "src/icons";
 
-type Trend = "stable" | "increasing" | "decreasing";
 const iconName = {
-  stable: "minus",
-  increasing: "arrowup",
-  decreasing: "arrowdown",
+  [Trend.Stable]: null,
+  [Trend.Up]: "arrowup",
+  [Trend.Down]: "arrowdown",
 } as const;
 export const TrendIcon: React.FC<
-  { trend: Trend } & Omit<ComponentProps<typeof Icon>, "name">
+  { trend: Trend | undefined | null } & Omit<
+    ComponentProps<typeof Icon>,
+    "name"
+  >
 > = ({ trend, ...props }) => {
-  const icon = iconName[trend];
-  return <Icon {...props} name={icon} />;
+  const icon = trend ? iconName[trend] : undefined;
+  return icon ? <Icon {...props} name={icon} /> : null;
 };